Don't dispose of linked assembly in VerifyLinkingOfOtherAssemblies
Disposing of the assembly leads to code after it to throw if it triggers a deferred read of something in the assembly.  This wasn't causing any issues in the illinker test framework.  I'm making changes in Unity and I ran into this.

There is no need to dispose of the assembly at this point.  The resolver will be diposed and will handle disposing of the assembly.