Remove AnalysisContext and ScopeStack from method body scanner (#2747)
* Motivation:

For compiler-generated code dataflow, we may need to analyze a
compiler-generated method multiple times as part of the closure of
methods reached from user code. This is necessary because as we
analyze this closure of methods we might find new captured state that
flows into a previously analyzed method in the same closure.

To avoid spurious warnings, we need a deduplication mechanism similar
to what we do in the analyzer. The eventual goal is for the linker to
have something similar to the analyzer's TrimAnalysisPatternStore that
deduplicates warnings produced from multiple scans of the same method
body.

We also might need the ability to run the method body scanner without
producing any warnings, when analyzing compiler-generated methods
reachable by reflection. (The linker will warn about reflection over
compiler-generated code, and won't produce the full set of diagnostics
for these methods, but still needs to scan them to ensure dependencies
are kept).

The first step is to remove some levels of indirection around the
message origins, to make it easier to cache messages by origin in the
future.

* This change:

This is an initial factoring of ReflectionMethodBodyScanner that just
gets rid of AnalysisContext and ScopeStack usages in that class. No
functional changes are intended.

Some of the ScopeStack logic is a bit subtle: the shared global
ScopeStack meant that it was difficult to tell which scope might be
used when reporting a warning - it was path-dependent. I pushed all of
the ScopeStack.CurrentScope.Origin calls up into callsites outside of
ReflectionMethodBodyScanner, so that the scanner logic could avoid
concerning itself with pushing and popping scopes. Instead, it passes
message origins along explicitly when required.