Move shared dataflow value creation helpers out of HandleCallAction (#2757)
This builds on the work @jtschuster did in c2fff33 to share code for
the annotation checks.

I noticed that ReflectionMethodBodyScanner and HandleCallAction
started accumulating these helpers for creating dataflow values from
methods and I thought it would make sense for that to be handled as a
separate concern, so I moved them out to a separate class and adjusted
some of the callsites in ReflectionMethodBodyScanner to use the shared
version.

This was motivated in part by some of the factoring I am doing as a
follow-up to 137b7b9, but isn't directly related. (While working on
the follow-up, I came across a point where I needed to create a
dataflow value in shared code from outside HandleCallAction, and there
wasn't a very clean way to do this).