Don't depend on marking order in DAM type hierarchy (#2777)
This will allow ApplyDynamicallyAccessedMembersToTypeHierarchy to
be called from the scanner when it sees object.GetType to get
annotations on the type, without requiring that the type has been
marked yet. (It will be marked during the final pass after the
flow analysis has converged.)

This will still not put unannotated base types into the cache (I
didn't want to introduce another cache of every marked type) -
but it means that now we need to walk up the unannotated base
types. Whereas before, relying on marking order meant that a type
not in the cache could be assumed to have None annotations.