Track attributes through Roslyn-generated types (#2786)
Type parameters for Roslyn-generated types are all alpha renames, so
there should be a unique original type parameter to map back to, which
can be used to get the original user attributes applied to the type
parameter. This change tries to build a map back from types to
constructor calls, which should map to the original type parameter annotations.

Only handles async and iterator state machines fully right now. Full support for
lambdas and local functions is left as future work.