Change the checks to only throw if we see a truly unexpected combination
I scoped the exception down to the case of a typed value which is of interesting type. Currently we don't think this can every happen and it's basically the only case where there's a potential for analysis hole.

I removed the new warning as there's nothing to warn about really. Originally we had the default case warn, but I don't think that makes sense currently.

I also added a test case which actually reproes the failure from the runtime - not exactly (it seems the compiler in runtime is slightly different or something else is going on and the exact same code from runtime produces different IL in the linker tests), but modified so that it does hit the exception in unchanged code.

I also validated that with this change the linker doesn't fail on runtime anymore.