Add two specific tests for type hierarchy marking (#2814)
First one is for the case where Derived type is annotated and has a virtual RUC method (and the base type has that same method with RUC as well). The warning should be generated for the base RUC method.

Second one is for type hierarchy marking of non-public members on base types when the derived type has the annotation. This is a test for https://github.com/dotnet/linker/issues/2813