Move the ILLink targets from the SDK repo to the linker repo
In general, a task invocation in MSBuild and the task definition are
inseparable. Adding parameters to the task will either result in errors
if the target is updated and too new, or silently skipping the parameter
if the target is too old.

Best practice is to ship both as a pair to prevent this problem. Testing is
harder, but we can add some limited testing if necessary to the linker repo.