Analyze nested functions in Roslyn analyzer (#2892)
This is the Roslyn analyzer counterpart to
bc46e445deb1411cc597019d693ddc5b4e5e24f4.

It adds detection of captured locals, and conservatively analyzes
them with all values that it encounters in the method.  The
analyzer makes no attempt to warn on reflection access to
compiler-generated fields, because the compilation strategy for
these is not known by the analyzer.

Unlike the linker, this doesn't do conservative analysis for
hoisted locals in state machine methods. The analyzer already
understands control flow within state machines the same way it
does for normal methods, so we just continue using that behavior
here. We could consider improving the linker's analysis to match,
but it would require a deeper understanding of the state machine
methods.