Add more tests for parameter dataflow (#2918)
Add tests for some edge cases due to the analyzer not treating
captured parameters as hoisted variables like the linker does.

Currently the analyzer doesn't detect hoisted parameters, and it
will treat them like normal parameters - warning if a parameter
is assigned a value with an annotation that doesn't match the
parameter annotation. The linker will treat captured parameters
the same as other captured variables, tracking all values that
get assigned. So there is a subtle difference in the warning
behavior:

- Linker won't warn on assignment to annotated captured
  parameter, but analyzer will.

- Linker will produce dataflow warnings for all values assigned
  to a captured parameter, but analyzer will not.