New constraint on a generic argument involved in MakeGenericType/Method should warn (#2942)
For correctness linker must be able to validate that generic parameters with new constraint will have their default .ctor preserved. If that can't be guaranteed, it needs to warn.

Added tests for cases around this.

Nullable<> is the only exception to this rule, so changed the code to completely skip any validation for it.