Build with preview5 SDK (#2954)
dede5f58dc9b7296de884ed62688b35d5b154b78 didn't fix the dependency
flow, because the packages for some reason are still being produced
with a preview7 in the runtimeconfig. Lacking better workarounds, this
change builds with the preview5 SDK that was selected in #2884 before
we decided to update to preview7 (due to a dotnet-format bug). This
disables the lint job to prevent those failures. We should turn it
back on once we can use preview7 (once dotnet/runtime and dotnet/sdk
are on preview7).