Revert analyzer to lower CodeAnalysis version (#2968)
Analyzers look for a newer version of Microsoft.CodeAnalysis than exists in the latest version of Visual Studio, and so they don't run in FullFramework tests in dotnet/sdk#26993. This reverts the version used in the analyzers.

PR #2969 will undo this change once Visual Studio ships with the correct version of Microsoft.CodeAnalysis.