Fix exception when generating warning for method definition (#2972)
If we produce a warning for the method itself (no IL offset) and the method has debug symbols, it can happen that there's no sequence point for the first instruction. In that case the current code will crash because we expect to always find sequence point for offset 0.

Fix this by looking for the first sequence point instead.