Upgrade MSBuild ref and downgrade global.json
The official build has an old version of MSBuild so we can't use
a new version of MSBuild. However, we do need a higher version of MSBuild
as our ref for the tasks, so this also upgrades the version of MSBuild.

The two changes are technically separate, even though they both involve
MSBuild.