Rework the change to guarantee that all accessed to Body are after the constant prop/branch removal happened on the method.
This does have one possibly negative impact: the issue described in https://github.com/dotnet/linker/issues/2937 is now consistent and happens always.

Basically if there's a local function which is going to be removed due to branch removal and if the body of that method contains code which produces a warning due to generic parameter validation, such warning will always be generated even though it's "dead" code and even if it's suppressed via RUC or similar.

Basically in such case the analysis can't figure out to which method the local function belongs (since the call site has been removed).