Improves GetDisplayName for types (#3134)
If there's a TypeReference which is not a definition (for example it's a by-ref, or a pointer, ...) and it's also a nested type, the current code would not include the owning type in the type name. For type defs we always include the owning type name.

This change fixes that.

Also a small tweak in the test infra to avoid adding the same values into a hashset every time we visit a type.