Warn on reflection access to instance methods on RUC types (#3147)
Fixes an analysis hole where reflecting on instance methods of RUC
types produced no warnings. This isn't trim-compatible in case the
methods are decorated with attributes with RUC constructors, or in
case they are invoked on a null instance. We don't warn for RUC
attributes on methods in RUC types, but instead prevent reflection
access from creating these attributes by warning on the reflection
access to methods (including instance methods).

This leaves the attribute warnings for fields as-is - so instance
fields will get attribute warnings unlike instance methods, because we
don't warn on reflection access to instance fields in RUC types. This
is currently inconsistent with the analyzer behavior, which doesn't warn
for attributes in this case.